Fixing bug with TXMLOdb on macOS
When trying to sort a run on macOS, grsisort crashed because it failed
to read the ODB correctly. After some digging and comparing outputs
between running grsisort on Ubuntu vs macOS, I found that apparently
there are some null-characters at the end of the buffer provided to the
constructor. Changing the size to ignore all of these fixed this bug.

The additional null-characters might come from byte-padding when the ODB
information is written, at least in the run I looked at for this, the
ODB information ended in the middle of a line with zeroes padding out
the rest of the 16-byte line. And the ODB written at the very end of the
file also has a line-feed (0x0a) followed by a null-character (0x00) at
the very end.

I don't fully understand why this is only an issue on macOS and not on
linux, but the changed code also works on linux.

Also fixed the throwing of exceptions (instead of just creating
exceptions) when encountering an issue like this.